fix improper mExecStopReq and bad entrance 6-tower

+++ include/d_actor.hpp
@@ -5,5 +5,5 @@
 class dActor_c
 {
 public:
-    static u32 mExecStopReq;
+    static u8 mExecStopReq;
 };
